# Performance of transform! on SubDataFrame