--- a/src/abstractdataframe/selection.jl
+++ b/src/abstractdataframe/selection.jl
@@ -969,7 +969,8 @@ $TRANSFORMATION_COMMON_RULES
 
 See [`select`](@ref) for examples.
 """
-transform!(df::DataFrame, @nospecialize(args...); renamecols::Bool=true, threads::Bool=true) =
+transform!(df::DataFrame, @nospecialize(args...);
+           renamecols::Bool=true, threads::Bool=true) =
     _replace_columns!(df, select(df, :, args..., copycols=false, renamecols=renamecols, threads=threads))
 
 transform!(df::SubDataFrame, @nospecialize(args...); renamecols::Bool=true, threads::Bool=true) =