--- a/src/dataframe/dataframe.jl
+++ b/src/dataframe/dataframe.jl
@@ -1812,9 +1812,8 @@ end
 
 # This is not exactly copy! as in general we allow axes to be different
 function _replace_columns!(df::DataFrame, newdf::DataFrame)
-    # here we do not support wastransform argument like for SubDataFrame
-    # because by default in DataFrame case columns are not copied
-    # so we need to pass `:` to select! to handle this case correctly
+    # for DataFrame object here we do not support keep_present keyword argument
+    # like for SubDataFrame because here transform! always falls back to select!
     @assert ncol(newdf) == 0 || nrow(df) == nrow(newdf)
     copy!(_columns(df), _columns(newdf))
     copy!(_names(index(df)), _names(newdf))