Fixes https://github.com/JuliaData/DataFrames.jl/issues/3069.

What this PR addresses.

1. The initially reported performance issue. After the PR:
```
julia> using DataFrames, BenchmarkTools

julia> df = DataFrame(rand(100000, 100), :auto);

julia> subdf = view(df, df.x1 .>= 0.5, :);

julia> @btime transform!($subdf, :x1 => maximum => :x1);
  289.800 μs (128 allocations: 1.15 MiB)

julia> gdf = groupby(subdf, :x2);

julia> @btime transform!($gdf, :x1 => maximum => :x1);
  1.090 ms (300 allocations: 4.21 MiB)
```
and before the PR:
```
julia> using DataFrames, BenchmarkTools

julia> df = DataFrame(rand(100000, 100), :auto);

julia> subdf = view(df, df.x1 .>= 0.5, :);

julia> @btime transform!($subdf, :x1 => maximum => :x1);
  37.446 ms (1642 allocations: 114.68 MiB)

julia> gdf = groupby(subdf, :x2);

julia> @btime transform!($gdf, :x1 => maximum => :x1);
  35.303 ms (4853 allocations: 117.23 MiB)
```

2. Aliasing issue in `select[!]` and `transform[!]` on `GroupedDataFrame`

After the PR:
```
julia> df = DataFrame(a=1:2, b=3:4);

julia> gdf = groupby(df, :a);

julia> res = select(gdf, :b, :a => :c, :b => :d, copycols=false);

julia> res.a === res.c
false

julia> res.b === res.d
false
```
and before the PR:
```
julia> df = DataFrame(a=1:2, b=3:4);

julia> gdf = groupby(df, :a);

julia> res = select(gdf, :b, :a => :c, :b => :d, copycols=false);

julia> res.a === res.c
true

julia> res.b === res.d
true
```