@bkamins 2022-06-08
> Is there a performance impact for transform when the number of columns is large?

It is negligible, for 10,000 columns it is around 500μs:
```
julia> df = DataFrame(rand(100, 10000), :auto);

julia> using BenchmarkTools

julia> @benchmark _dealias_dataframe($df)
BenchmarkTools.Trial: 9883 samples with 1 evaluation.
 Range (min … max):  432.800 μs …   4.738 ms  ┊ GC (min … max): 0.00% … 82.65%
 Time  (median):     485.300 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   503.228 μs ± 112.816 μs  ┊ GC (mean ± σ):  1.19% ±  4.68%

   ▁▅ ▁▄▃█▇▁
  ▂██▇██████▇▅▄▄▄▅▄▄▃▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  433 μs           Histogram: frequency by time          853 μs <

 Memory estimate: 344.42 KiB, allocs estimate: 11.
```

> In theory, couldn't the check be restricted to columns that have been used as inputs of transformations?

It could, and this is what we do for `DataFrame` in `select` and `transform`. But I have reviewed the code before proposing this approach and decided that it is overly complex. Actually I wish I had this (i.e. using `IdDict`) idea earlier as this approach is much more robust than checking input-output mapping. Still for now I decided to leave things as they are in the old code as processing `DataFrame` fast is more important since the "core" of the operation is easier.